refactor: split payroll loan car routes
