+++ HRPayMaster/server/routes.ts
@@ -6,11 +6,7 @@ import { storage } from "./storage";
 import { 
   insertDepartmentSchema,
   insertEmployeeSchema,
-  insertPayrollRunSchema,
-  insertPayrollEntrySchema,
   insertVacationRequestSchema,
-  insertLoanSchema,
-  insertCarSchema,
   insertCarAssignmentSchema,
   insertNotificationSchema,
   insertEmailAlertSchema,
@@ -176,268 +172,6 @@ export async function registerRoutes(app: Express): Promise<Server> {
     }
   });
 
-  // Payroll routes
-  app.get("/api/payroll", async (req, res, next) => {
-    try {
-      const payrollRuns = await storage.getPayrollRuns();
-      res.json(payrollRuns);
-    } catch (error) {
-      next(new HttpError(500, "Failed to fetch payroll runs"));
-    }
-  });
-
-  app.get("/api/payroll/:id", async (req, res, next) => {
-    try {
-      const payrollRun = await storage.getPayrollRun(req.params.id);
-      if (!payrollRun) {
-        return next(new HttpError(404, "Payroll run not found"));
-      }
-      res.json(payrollRun);
-    } catch (error) {
-      next(new HttpError(500, "Failed to fetch payroll run"));
-    }
-  });
-
-  app.post("/api/payroll", async (req, res, next) => {
-    try {
-      const payrollRun = insertPayrollRunSchema.parse(req.body);
-      const newPayrollRun = await storage.createPayrollRun(payrollRun);
-      res.status(201).json(newPayrollRun);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return next(new HttpError(400, "Invalid payroll data", error.errors));
-      }
-      next(new HttpError(500, "Failed to create payroll run"));
-    }
-  });
-
-  app.post("/api/payroll/generate", async (req, res, next) => {
-    try {
-      const { period, startDate, endDate } = req.body;
-      
-      if (!period || !startDate || !endDate) {
-        return next(new HttpError(400, "Period, start date, and end date are required"));
-      }
-
-      // Get all active employees
-      const employees = await storage.getEmployees();
-      const activeEmployees = employees.filter(emp => emp.status === "active");
-
-      if (activeEmployees.length === 0) {
-        return next(new HttpError(400, "No active employees found"));
-      }
-
-      // Get loans, vacation requests, and employee events for the period
-      const loans = await storage.getLoans();
-      const vacationRequests = await storage.getVacationRequests();
-      const employeeEvents = await storage.getEmployeeEvents();
-      
-      // Calculate working days in the period
-      const start = new Date(startDate);
-      const end = new Date(endDate);
-      const totalDays = Math.ceil((end.getTime() - start.getTime()) / (1000 * 60 * 60 * 24)) + 1;
-      const workingDays = Math.max(1, totalDays - Math.floor(totalDays / 7) * 2); // Approximate working days
-
-      // Calculate totals
-      let grossAmount = 0;
-      let totalDeductions = 0;
-
-      const payrollEntries = await Promise.all(activeEmployees.map(async employee => {
-        const monthlySalary = parseFloat(employee.salary);
-        
-        // Calculate vacation days for this employee in the period
-        const employeeVacations = vacationRequests.filter(v => 
-          v.employeeId === employee.id && 
-          v.status === "approved" &&
-          new Date(v.startDate) <= end && 
-          new Date(v.endDate) >= start
-        );
-        
-        const vacationDays = employeeVacations.reduce((total, vacation) => {
-          const vacStart = new Date(Math.max(new Date(vacation.startDate).getTime(), start.getTime()));
-          const vacEnd = new Date(Math.min(new Date(vacation.endDate).getTime(), end.getTime()));
-          return total + Math.ceil((vacEnd.getTime() - vacStart.getTime()) / (1000 * 60 * 60 * 24)) + 1;
-        }, 0);
-
-        // Use employee's standard working days (default 26 if not set)
-        const employeeWorkingDays = employee.standardWorkingDays || 26;
-        
-        // Calculate actual working days for this employee
-        const actualWorkingDays = Math.max(0, employeeWorkingDays - vacationDays);
-        
-        // Calculate pro-rated salary based on employee's standard working days
-        const baseSalary = employee.status === "active" ? 
-          (monthlySalary * actualWorkingDays / employeeWorkingDays) : 0;
-
-        // Calculate loan deductions for this employee
-        const employeeLoans = loans.filter(l => 
-          l.employeeId === employee.id && 
-          l.status === "active" &&
-          parseFloat(l.remainingAmount) > 0
-        );
-        
-        const loanDeduction = employeeLoans.reduce((total, loan) => {
-          return total + Math.min(parseFloat(loan.monthlyDeduction), parseFloat(loan.remainingAmount));
-        }, 0);
-
-        // Calculate employee events (bonuses, deductions, etc.) for this period
-        const periodStart = new Date(startDate);
-        const periodEnd = new Date(endDate);
-        
-        const employeeEventsInPeriod = employeeEvents.filter(event => 
-          event.employeeId === employee.id && 
-          event.affectsPayroll &&
-          event.status === "active" &&
-          new Date(event.eventDate) >= periodStart && 
-          new Date(event.eventDate) <= periodEnd
-        );
-
-        const bonusAmount = employeeEventsInPeriod
-          .filter(event => ['bonus', 'allowance', 'overtime'].includes(event.eventType))
-          .reduce((total, event) => total + parseFloat(event.amount), 0);
-
-        const eventDeductions = employeeEventsInPeriod
-          .filter(event => ['deduction', 'penalty'].includes(event.eventType))
-          .reduce((total, event) => total + parseFloat(event.amount), 0);
-
-        // Add bonuses to get gross pay
-        const grossPay = baseSalary + bonusAmount;
-
-        // Calculate standard deductions (optional - can be configured per company)
-        const taxDeduction = 0; // No automatic tax deduction
-        const socialSecurityDeduction = 0; // No automatic social security deduction
-        const healthInsuranceDeduction = 0; // No automatic health insurance deduction
-
-        const otherDeductions = eventDeductions;
-        
-        const totalEmpDeductions = taxDeduction + socialSecurityDeduction + healthInsuranceDeduction + loanDeduction + otherDeductions;
-        const netPay = Math.max(0, grossPay - totalEmpDeductions);
-
-        grossAmount += grossPay;
-        totalDeductions += totalEmpDeductions;
-
-        // Create notifications for significant payroll events
-        let adjustmentReason = "";
-        if (vacationDays > 0) {
-          adjustmentReason += `${vacationDays} vacation days. `;
-          
-          // Create notification for vacation impact
-          await storage.createNotification({
-            employeeId: employee.id,
-            type: "vacation_approved",
-            title: "Vacation Deduction Applied",
-            message: `${vacationDays} vacation days deducted from ${period} payroll`,
-            priority: "medium",
-            status: "unread",
-            expiryDate: endDate,
-            daysUntilExpiry: 0,
-            emailSent: false
-          });
-        }
-        
-        if (loanDeduction > 0) {
-          adjustmentReason += `Loan deduction: ${loanDeduction.toFixed(2)} KWD. `;
-          
-          // Create notification for loan deduction
-          await storage.createNotification({
-            employeeId: employee.id,
-            type: "loan_deduction",
-            title: "Loan Deduction Applied",
-            message: `${loanDeduction.toFixed(2)} KWD deducted for loan repayment in ${period}`,
-            priority: "low",
-            status: "unread",
-            expiryDate: endDate,
-            daysUntilExpiry: 0,
-            emailSent: false
-          });
-        }
-
-        return {
-          employeeId: employee.id,
-          grossPay: grossPay.toString(),
-          baseSalary: baseSalary.toString(),
-          bonusAmount: bonusAmount.toString(),
-          workingDays: employeeWorkingDays,
-          actualWorkingDays: actualWorkingDays,
-          vacationDays: vacationDays,
-          taxDeduction: taxDeduction.toString(),
-          socialSecurityDeduction: socialSecurityDeduction.toString(),
-          healthInsuranceDeduction: healthInsuranceDeduction.toString(),
-          loanDeduction: loanDeduction.toString(),
-          otherDeductions: otherDeductions.toString(),
-          netPay: netPay.toString(),
-          adjustmentReason: adjustmentReason.trim() || null,
-        };
-      }));
-
-      const netAmount = grossAmount - totalDeductions;
-
-      // Create payroll run
-      const payrollRun = await storage.createPayrollRun({
-        period,
-        startDate,
-        endDate,
-        grossAmount: grossAmount.toString(),
-        totalDeductions: totalDeductions.toString(),
-        netAmount: netAmount.toString(),
-        status: "completed"
-      });
-
-      // Create payroll entries
-      for (const entry of payrollEntries) {
-        await storage.createPayrollEntry({
-          ...entry,
-          payrollRunId: payrollRun.id
-        });
-      }
-
-      // Update loan remaining amounts
-      for (const loan of loans.filter(l => l.status === "active")) {
-        const loanDeduction = payrollEntries.find(entry => entry.employeeId === loan.employeeId)?.loanDeduction;
-        if (loanDeduction && parseFloat(loanDeduction) > 0) {
-          const newRemaining = Math.max(0, parseFloat(loan.remainingAmount) - parseFloat(loanDeduction));
-          await storage.updateLoan(loan.id, {
-            remainingAmount: newRemaining.toString(),
-            status: newRemaining <= 0 ? "completed" : "active"
-          });
-        }
-      }
-
-      res.status(201).json(payrollRun);
-    } catch (error) {
-      console.error("Payroll generation error:", error);
-      next(new HttpError(500, "Failed to generate payroll"));
-    }
-  });
-
-  app.put("/api/payroll/:id", async (req, res, next) => {
-    try {
-      const updates = insertPayrollRunSchema.partial().parse(req.body);
-      const updatedPayrollRun = await storage.updatePayrollRun(req.params.id, updates);
-      if (!updatedPayrollRun) {
-        return next(new HttpError(404, "Payroll run not found"));
-      }
-      res.json(updatedPayrollRun);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return next(new HttpError(400, "Invalid payroll data", error.errors));
-      }
-      next(new HttpError(500, "Failed to update payroll run"));
-    }
-  });
-
-  app.delete("/api/payroll/:id", async (req, res, next) => {
-    try {
-      const deleted = await storage.deletePayrollRun(req.params.id);
-      if (!deleted) {
-        return next(new HttpError(404, "Payroll run not found"));
-      }
-      res.status(204).send();
-    } catch (error) {
-      next(new HttpError(500, "Failed to delete payroll run"));
-    }
-  });
-
   // Dashboard stats route
   app.get("/api/dashboard/stats", async (req, res, next) => {
     try {
@@ -526,132 +260,6 @@ export async function registerRoutes(app: Express): Promise<Server> {
       res.status(204).send();
     } catch (error) {
       next(new HttpError(500, "Failed to delete vacation request"));
-    }
-  });
-
-  // Loan routes
-  app.get("/api/loans", async (req, res, next) => {
-    try {
-      const loans = await storage.getLoans();
-      res.json(loans);
-    } catch (error) {
-      next(new HttpError(500, "Failed to fetch loans"));
-    }
-  });
-
-  app.get("/api/loans/:id", async (req, res, next) => {
-    try {
-      const loan = await storage.getLoan(req.params.id);
-      if (!loan) {
-        return next(new HttpError(404, "Loan not found"));
-      }
-      res.json(loan);
-    } catch (error) {
-      next(new HttpError(500, "Failed to fetch loan"));
-    }
-  });
-
-  app.post("/api/loans", async (req, res, next) => {
-    try {
-      const loan = insertLoanSchema.parse(req.body);
-      const newLoan = await storage.createLoan(loan);
-      res.status(201).json(newLoan);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return next(new HttpError(400, "Invalid loan data", error.errors));
-      }
-      next(new HttpError(500, "Failed to create loan"));
-    }
-  });
-
-  app.put("/api/loans/:id", async (req, res, next) => {
-    try {
-      const updates = insertLoanSchema.partial().parse(req.body);
-      const updatedLoan = await storage.updateLoan(req.params.id, updates);
-      if (!updatedLoan) {
-        return next(new HttpError(404, "Loan not found"));
-      }
-      res.json(updatedLoan);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return next(new HttpError(400, "Invalid loan data", error.errors));
-      }
-      next(new HttpError(500, "Failed to update loan"));
-    }
-  });
-
-  app.delete("/api/loans/:id", async (req, res, next) => {
-    try {
-      const deleted = await storage.deleteLoan(req.params.id);
-      if (!deleted) {
-        return next(new HttpError(404, "Loan not found"));
-      }
-      res.status(204).send();
-    } catch (error) {
-      next(new HttpError(500, "Failed to delete loan"));
-    }
-  });
-
-  // Car routes
-  app.get("/api/cars", async (req, res, next) => {
-    try {
-      const cars = await storage.getCars();
-      res.json(cars);
-    } catch (error) {
-      next(new HttpError(500, "Failed to fetch cars"));
-    }
-  });
-
-  app.get("/api/cars/:id", async (req, res, next) => {
-    try {
-      const car = await storage.getCar(req.params.id);
-      if (!car) {
-        return next(new HttpError(404, "Car not found"));
-      }
-      res.json(car);
-    } catch (error) {
-      next(new HttpError(500, "Failed to fetch car"));
-    }
-  });
-
-  app.post("/api/cars", async (req, res, next) => {
-    try {
-      const car = insertCarSchema.parse(req.body);
-      const newCar = await storage.createCar(car);
-      res.status(201).json(newCar);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return next(new HttpError(400, "Invalid car data", error.errors));
-      }
-      next(new HttpError(500, "Failed to create car"));
-    }
-  });
-
-  app.put("/api/cars/:id", async (req, res, next) => {
-    try {
-      const updates = insertCarSchema.partial().parse(req.body);
-      const updatedCar = await storage.updateCar(req.params.id, updates);
-      if (!updatedCar) {
-        return next(new HttpError(404, "Car not found"));
-      }
-      res.json(updatedCar);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return next(new HttpError(400, "Invalid car data", error.errors));
-      }
-      next(new HttpError(500, "Failed to update car"));
-    }
-  });
-
-  app.delete("/api/cars/:id", async (req, res, next) => {
-    try {
-      const deleted = await storage.deleteCar(req.params.id);
-      if (!deleted) {
-        return next(new HttpError(404, "Car not found"));
-      }
-      res.status(204).send();
-    } catch (error) {
-      next(new HttpError(500, "Failed to delete car"));
     }
   });
 
@@ -1051,23 +659,6 @@ export async function registerRoutes(app: Express): Promise<Server> {
         return next(new HttpError(400, "Invalid query parameters", error.errors));
       }
       next(new HttpError(500, "Failed to fetch employee report"));
-    }
-  });
-
-  // Payroll entry routes
-  app.put("/api/payroll/entries/:id", async (req, res, next) => {
-    try {
-      const updates = insertPayrollEntrySchema.partial().parse(req.body);
-      const updatedEntry = await storage.updatePayrollEntry(req.params.id, updates);
-      if (!updatedEntry) {
-        return next(new HttpError(404, "Payroll entry not found"));
-      }
-      res.json(updatedEntry);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return next(new HttpError(400, "Invalid payroll entry data", error.errors));
-      }
-      next(new HttpError(500, "Failed to update payroll entry"));
     }
   });
